+++ src/OneBlinkFormBase.tsx
@@ -1006,6 +1006,7 @@ function OneBlinkFormBase({
                                       {visiblePages.map(
                                         (
                                           pageElement: FormTypes.PageElement,
+                                          index,
                                         ) => (
                                           <PageFormElements
                                             key={pageElement.id}
@@ -1030,6 +1031,7 @@ function OneBlinkFormBase({
                                             setFormSubmission={
                                               setFormSubmission
                                             }
+                                            isFirstPage={index === 0}
                                           />
                                         ),
                                       )}